test: split admin analytics tests
