+++ crates/dcc-mcp-gateway/src/gateway/admin/tests.rs
@@ -3227,147 +3227,6 @@ webhooks:
         assert_eq!(limited_body["total"], 1);
     }
 
-    fn analytics_audit_record(
-        request_id: &str,
-        action: &str,
-        instance_id: Option<&str>,
-        agent_id: Option<&str>,
-        success: bool,
-    ) -> AdminAuditRecord {
-        AdminAuditRecord {
-            timestamp: std::time::SystemTime::now()
-                .checked_sub(Duration::from_secs(60))
-                .unwrap(),
-            request_id: request_id.to_string(),
-            trace_id: Some(format!("trace-{request_id}")),
-            span_id: None,
-            parent_span_id: None,
-            method: Some("tools/call".to_string()),
-            instance_id: instance_id.map(str::to_string),
-            session_id: None,
-            transport: Some("rest".to_string()),
-            agent_id: agent_id.map(str::to_string),
-            agent_name: agent_id.map(|id| format!("Agent {id}")),
-            agent_model: None,
-            actor_id: None,
-            actor_name: None,
-            actor_email_hash: None,
-            client_platform: None,
-            client_os: None,
-            client_host: None,
-            auth_subject: None,
-            source_ip: None,
-            attribution_trust: None,
-            parent_request_id: None,
-            action: action.to_string(),
-            dcc_type: Some("maya".to_string()),
-            success,
-            error: (!success).then(|| "boom".to_string()),
-            duration_ms: Some(42),
-            token_accounting: Some(token_telemetry("json", 100, 40)),
-            llm_usage: None,
-        }
-    }
-
-    #[tokio::test]
-    async fn test_admin_analytics_overview_returns_unique_counts() {
-        let audit: AuditLog = Mutex::new(vec![
-            analytics_audit_record(
-                "analytics-1",
-                "maya.inst-a.scene__info",
-                Some("inst-a"),
-                Some("agent-a"),
-                true,
-            ),
-            analytics_audit_record(
-                "analytics-2",
-                "maya.inst-b.scene__info",
-                Some("inst-b"),
-                Some("agent-a"),
-                false,
-            ),
-            analytics_audit_record(
-                "analytics-3",
-                "maya.inst-b.scene__info",
-                Some("inst-b"),
-                None,
-                true,
-            ),
-        ]);
-        let state = AdminState::new(make_gateway_state()).with_audit_log(Arc::new(audit));
-        let router = build_admin_router(state);
-
-        let (status, body) = body_json(router, "/api/analytics/overview?range=7d").await;
-
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(body["kpi"]["calls_total"], 3);
-        assert_eq!(body["kpi"]["calls_failed"], 1);
-        assert_eq!(body["kpi"]["unique_instances"], 2);
-        assert_eq!(body["kpi"]["unique_agents"], 1);
-        assert_eq!(body["daily_series"][0]["max_duration_ms"], 42);
-    }
-
-    #[tokio::test]
-    async fn test_admin_analytics_csv_export_escapes_cells() {
-        let audit: AuditLog = Mutex::new(vec![analytics_audit_record(
-            "analytics-csv-1",
-            "=cmd,tool\nline",
-            Some("inst-a"),
-            Some("@agent-a"),
-            true,
-        )]);
-        let state = AdminState::new(make_gateway_state()).with_audit_log(Arc::new(audit));
-        let router = build_admin_router(state);
-
-        let (status, _headers, body) = body_text_with_accept(
-            router,
-            "/api/analytics/export?range=7d&format=csv",
-            "text/csv",
-        )
-        .await;
-
-        assert_eq!(status, StatusCode::OK);
-        assert!(body.starts_with("request_id,timestamp,action,dcc_type"));
-        assert!(
-            body.contains("\"'=cmd,tool\nline\""),
-            "expected formula-prefixed and quoted action cell, got:\n{body}"
-        );
-        assert!(
-            body.contains("'@agent-a"),
-            "expected formula-prefixed agent id, got:\n{body}"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_v1_debug_analytics_export_mirrors_admin_export() {
-        let audit: AuditLog = Mutex::new(vec![analytics_audit_record(
-            "analytics-v1-csv-1",
-            "maya.inst.scene__info",
-            Some("inst-a"),
-            Some("agent-a"),
-            true,
-        )]);
-        let state = AdminState::new(make_gateway_state()).with_audit_log(Arc::new(audit));
-        let router = build_v1_debug_router(state);
-
-        let (status, headers, body) = body_text_with_accept(
-            router,
-            "/v1/debug/analytics/export?range=7d&format=csv",
-            "text/csv",
-        )
-        .await;
-
-        assert_eq!(status, StatusCode::OK);
-        assert!(
-            headers
-                .get(header::CONTENT_TYPE)
-                .and_then(|value| value.to_str().ok())
-                .is_some_and(|value| value.starts_with("text/csv")),
-        );
-        assert!(body.starts_with("request_id,timestamp,action,dcc_type"));
-        assert!(body.contains("analytics-v1-csv-1"));
-    }
-
     // ── unknown routes ────────────────────────────────────────────────────
 
     #[tokio::test]
